Move the damage functionality under battle
Since most situation won't require the damage calculator itself, it
should be part of the battle functionality.
